Remove Deprecated API calls (#146)
* ⚰️ Remove deprecated API calls

Removed references to Resource Navigator (which ws deprecated for a long
time and has since been removed from the Eclipse API.

Replaced addFastView() with calls to addShowViewShortcut().

* ⚰️ Removed entries related to CVS

CVS was removed in Eclipse 2022-03.

* 💡 Clarified some comments

Added an entry for where the ShowViewShortcuts end up in the interface
so that the next person can find them.

Fixes #127